requirements.txt
requests==2.22.0
pycrypto==2.6.1
